node: use `uint64_t` for dbcache byte constants
The byte-unit literals used by the dbcache policy are `uint64_t`, but several dbcache-local constants still store those values as `size_t`.

Move the min, default, max, and high-default byte constants to `uint64_t` while keeping `GetDefaultDBCache()` and the cache-splitting API at their existing `size_t` boundaries.
With `MAX_DBCACHE_BYTES` now matching the parsed byte count type, `CalculateDbCacheBytes()` no longer needs an explicit `std::min<uint64_t>`.

+++ src/node/dbcache.cpp
@@ -11,7 +11,7 @@
 #include <cstdint>
 
 //! Larger default dbcache on 64-bit systems with enough RAM.
-static constexpr size_t HIGH_DEFAULT_DBCACHE{1_GiB};
+static constexpr uint64_t HIGH_DEFAULT_DBCACHE{1_GiB};
 //! Minimum detected RAM required for HIGH_DEFAULT_DBCACHE.
 static constexpr uint64_t HIGH_DEFAULT_DBCACHE_MIN_TOTAL_RAM{4_GiB};
